Finish to implement factory and add various command line to test it

+++ Source/Flare/Game/FlareCompany.h
@@ -306,7 +306,7 @@ public:
 		return CompanyDescription->CustomizationPatternIndex;
 	}
 
-	inline int32 GetMoney() const
+	inline uint64 GetMoney() const
 	{
 		return CompanyData.Money;
 	}
